+++ Search/Test.hs
@@ -29,24 +29,24 @@ monads       = [list, may, io, eithererr, tcrn, ghc, mm]
 traversables = [list, may, eithererr, messagesprod]
 applicatives = [list, may, io, eithererr, tcrn, ghc, mm]
 
-joint f   = Trans { from = [f,f], to = [f], name = "join" }
-returnt f   = Trans { from = [], to = [f], name = "return" }
-sequ t f = Trans {from = [t, f], to = [f, t], name = "sequence" }
+joint f   = Trans { from = [f,f], to = [f], name = Simple "join" }
+returnt f   = Trans { from = [], to = [f], name = Simple "return" }
+sequ t f = Trans {from = [t, f], to = [f, t], name = Simple "sequence" }
 
 transes :: [Trans F]
 transes =
   concatMap (\x -> [joint x {-, returnt x -}]) monads
   ++ liftA2 sequ traversables applicatives
   ++
-  [ Trans { from = [list, may], to = [list], name = "catMaybes" }
-  , Trans { from = [ghc], to = [io], name = "runGhc _" }
-  , Trans { from = [mm], to = [eithererr, ghc], name = "runErrorT" }
-  , Trans { from = [tcrn], to = [io, messagesprod, may], name = "runTcInteractive _" }
-  , Trans { from = [messagesprod], to = [], name = "snd" }
-  , Trans { from = [may], to = [mm], name = "maybeErr" }
-  , Trans { from = [list], to = [mm], name = "headErr" }
-  , Trans { from = [tcrn], to = [mm], name = "inHoleEnv _ _" }
-  , Trans { from = [eithererr], to = [may], name = "either (const Nothing) Just" }
+  [ Trans { from = [list, may], to = [list], name = Simple "catMaybes" }
+  , Trans { from = [ghc], to = [io], name = Compound "runGhc _" }
+  , Trans { from = [mm], to = [eithererr, ghc], name = Simple "runErrorT" }
+  , Trans { from = [tcrn], to = [io, messagesprod, may], name = Compound "runTcInteractive _" }
+  , Trans { from = [messagesprod], to = [], name = Simple "snd" }
+  , Trans { from = [may], to = [mm], name = Simple "maybeErr" }
+  , Trans { from = [list], to = [mm], name = Simple "headErr" }
+  , Trans { from = [tcrn], to = [mm], name = Compound "inHoleEnv _ _" }
+  , Trans { from = [eithererr], to = [may], name = Compound "either (const Nothing) Just" }
   ]
 
 main :: IO ()
