+++ graphs/out.html
@@ -4,7 +4,7 @@
 .network { width: 500px; height:500px; }
 </style>
 <script type='text/javascript' src='vis.min.js'></script>
-<script type='text/javascript'>function draw(){network = new vis.Network(document.getElementById("net0"),{"edges":[{"style":"arrow","to":"real2","from":"real1","label":"IO"},{"style":"arrow","to":"real3","from":"real1","label":"(,) ErrUtils.Mesages"},{"style":"arrow","to":"real3","from":"real1","label":"Maybe"},{"style":"arrow","to":"out0","from":"real2","label":"IO"},{"style":"arrow","to":"real4","from":"real2","label":"List"},{"style":"arrow","to":"real4","from":"real3","label":"Maybe"},{"style":"arrow","to":"real5","from":"real3","label":"(,) ErrUtils.Mesages"},{"style":"arrow","to":"out1","from":"real4","label":"List"},{"style":"arrow","to":"real2","from":"in0","label":"List"}],"nodes":[{"color":"green","id":"in0","label":""},{"color":"green","id":"in1","label":""},{"color":"green","id":"in2","label":""},{"color":"red","id":"out0","label":""},{"color":"red","id":"out1","label":""},{"color":"gray","id":"real1","label":"runTcInteractive _"},{"color":"gray","id":"real2","label":"sequence"},{"color":"gray","id":"real3","label":"sequence"},{"color":"gray","id":"real4","label":"catMaybes"},{"color":"gray","id":"real5","label":"snd"}]},{});
+<script type='text/javascript'>function draw(){network = new vis.Network(document.getElementById("net0"),{"edges":[{"style":"arrow","to":"real1","from":"real0","label":"([], [])"},{"style":"arrow","to":"out1","from":"real0","label":"(Data.Maybe.Maybe, [])"},{"style":"arrow","to":"out0","from":"real1","label":"([], [])"},{"style":"arrow","to":"real0","from":"in0","label":"(Data.Maybe.Maybe, [])"},{"style":"arrow","to":"real0","from":"in1","label":"([], [])"}],"nodes":[{"color":"green","id":"in0","label":""},{"color":"green","id":"in1","label":""},{"color":"red","id":"out0","label":""},{"color":"red","id":"out1","label":""},{"color":"gray","id":"real0","label":"sequenceA"},{"color":"gray","id":"real1","label":"cycle"}]},{});
 }</script>
 </head>
 <body onload='draw()'>
